refactor(menu): ♻️ move navigation menu decoration out of the hooks class
The navigation pipeline was never dispatched by the hook system: no
hook fires late enough to see every extension's changes, so the skin
calls it directly. It now lives in its own class and returns the
merged notification data, so the skin no longer has a setter that is
only valid mid-render and the two classes no longer import each other.

The transformer takes SpecialPageFactory rather than reaching for
SpecialPage statically, which is what lets its tests run as unit
tests. The service is appended to the end of the skin's service list
so every existing constructor position is preserved. Note that
getTitleForAlias() returns null for an unregistered page where
getTitleFor() minted a Title unconditionally; neither caller can
reach that, and the empty string is the deliberate answer if one ever
does.
