+++ tests/phpunit/integration/Hooks/SkinHooksTest.php
@@ -6,17 +6,13 @@ namespace MediaWiki\Skins\Citizen\Tests\Integration\Hooks;
 
 use MediaWiki\Config\HashConfig;
 use MediaWiki\MainConfigNames;
-use MediaWiki\Message\Message;
 use MediaWiki\Output\OutputPage;
 use MediaWiki\Request\FauxRequest;
 use MediaWiki\ResourceLoader\Context;
 use MediaWiki\Skins\Citizen\Hooks\SkinHooks;
-use MediaWiki\Skins\Citizen\SkinCitizen;
-use MediaWiki\User\User;
 use MediaWikiIntegrationTestCase;
 use PHPUnit\Framework\MockObject\MockObject;
 use Skin;
-use SkinTemplate;
 
 /**
  * @group Citizen
@@ -97,212 +93,6 @@ class SkinHooksTest extends MediaWikiIntegrationTestCase {
 		$this->assertArrayNotHasKey( 'meta-viewport', $tags );
 	}
 
-	private function createSkinTemplateMock( string $skinName = 'citizen' ): SkinTemplate {
-		$user = $this->createMock( User::class );
-		$user->method( 'isRegistered' )->willReturn( true );
-		$user->method( 'isTemp' )->willReturn( false );
-
-		$sktemplate = $this->createMock( SkinTemplate::class );
-		$sktemplate->method( 'getSkinName' )->willReturn( $skinName );
-		$sktemplate->method( 'getUser' )->willReturn( $user );
-		$sktemplate->method( 'msg' )->willReturnCallback(
-			fn ( string $key ): Message => $this->createConfiguredMock( Message::class, [ 'text' => $key ] )
-		);
-
-		return $sktemplate;
-	}
-
-	public function testViewsMenuMapsIcons(): void {
-		$sktemplate = $this->createSkinTemplateMock();
-		$links = [
-			'views' => [
-				'view' => [ 'id' => 'ca-view' ],
-				'edit' => [ 'id' => 'ca-edit', 'class' => '' ],
-				'history' => [ 'id' => 'ca-history' ],
-			],
-		];
-
-		SkinHooks::onSkinTemplateNavigation( $sktemplate, $links );
-
-		$this->assertSame( 'eye', $links['views']['view']['icon'] );
-		$this->assertSame( 'edit', $links['views']['edit']['icon'] );
-		$this->assertSame( 'history', $links['views']['history']['icon'] );
-	}
-
-	public function testViewsMenuMergesVeAndSourceEdit(): void {
-		$sktemplate = $this->createSkinTemplateMock();
-		$links = [
-			'views' => [
-				've-edit' => [ 'id' => 'ca-ve-edit', 'class' => '' ],
-				'edit' => [ 'id' => 'ca-edit', 'class' => '' ],
-			],
-		];
-
-		SkinHooks::onSkinTemplateNavigation( $sktemplate, $links );
-
-		// When both exist, source edit should use wikiText icon
-		$this->assertSame( 'wikiText', $links['views']['edit']['icon'] );
-		$this->assertSame( 'edit', $links['views']['ve-edit']['icon'] );
-		// Both should have the merged CSS class
-		$this->assertStringContainsString( 'citizen-ve-edit-merged', $links['views']['ve-edit']['class'] );
-		$this->assertStringContainsString( 'citizen-ve-edit-merged', $links['views']['edit']['class'] );
-	}
-
-	public function testViewsMenuWithOnlySourceEdit(): void {
-		$sktemplate = $this->createSkinTemplateMock();
-		$links = [
-			'views' => [
-				'edit' => [ 'id' => 'ca-edit', 'class' => '' ],
-			],
-		];
-
-		SkinHooks::onSkinTemplateNavigation( $sktemplate, $links );
-
-		// Without VE, source edit should keep the edit icon
-		$this->assertSame( 'edit', $links['views']['edit']['icon'] );
-	}
-
-	public function testNavigationSkipsNonCitizen(): void {
-		$sktemplate = $this->createSkinTemplateMock( 'vector' );
-		$links = [
-			'views' => [
-				'edit' => [ 'id' => 'ca-edit', 'class' => '' ],
-			],
-		];
-
-		SkinHooks::onSkinTemplateNavigation( $sktemplate, $links );
-
-		$this->assertArrayNotHasKey( 'icon', $links['views']['edit'] );
-	}
-
-	public function testActionsMenuMapsIcons(): void {
-		$sktemplate = $this->createSkinTemplateMock();
-		$links = [
-			'actions' => [
-				'delete' => [ 'id' => 'ca-delete' ],
-				'move' => [ 'id' => 'ca-move' ],
-				'protect' => [ 'id' => 'ca-protect' ],
-			],
-		];
-
-		SkinHooks::onSkinTemplateNavigation( $sktemplate, $links );
-
-		$this->assertSame( 'trash', $links['actions']['delete']['icon'] );
-		$this->assertSame( 'move', $links['actions']['move']['icon'] );
-		$this->assertSame( 'lock', $links['actions']['protect']['icon'] );
-	}
-
-	public function testAssociatedPagesMenuMapsTalkPageIcons(): void {
-		$sktemplate = $this->createSkinTemplateMock();
-		$links = [
-			'associated-pages' => [
-				'main' => [ 'id' => 'ca-nstab-main' ],
-				'main_talk' => [ 'id' => 'ca-talk' ],
-			],
-		];
-
-		SkinHooks::onSkinTemplateNavigation( $sktemplate, $links );
-
-		$this->assertSame( 'speechBubbles', $links['associated-pages']['main_talk']['icon'] );
-		$this->assertSame( 'arrowPrevious', $links['associated-pages']['main']['icon'] );
-	}
-
-	public function testAssociatedPagesMenuMapsFileIcon(): void {
-		$sktemplate = $this->createSkinTemplateMock();
-		$links = [
-			'associated-pages' => [
-				'file' => [ 'id' => 'ca-nstab-image' ],
-			],
-		];
-
-		SkinHooks::onSkinTemplateNavigation( $sktemplate, $links );
-
-		$this->assertSame( 'image', $links['associated-pages']['file']['icon'] );
-	}
-
-	private function createSkinTemplateWithUser(
-		bool $isRegistered,
-		bool $isTemp
-	): SkinTemplate {
-		$user = $this->createMock( User::class );
-		$user->method( 'isRegistered' )->willReturn( $isRegistered );
-		$user->method( 'isTemp' )->willReturn( $isTemp );
-
-		$sktemplate = $this->createMock( SkinTemplate::class );
-		$sktemplate->method( 'getSkinName' )->willReturn( 'citizen' );
-		$sktemplate->method( 'getUser' )->willReturn( $user );
-
-		return $sktemplate;
-	}
-
-	public function testUserMenuRemovesUserpageForRegistered(): void {
-		$sktemplate = $this->createSkinTemplateWithUser( true, false );
-		$links = [
-			'user-menu' => [
-				'userpage' => [ 'id' => 'pt-userpage' ],
-				'preferences' => [ 'id' => 'pt-preferences' ],
-			],
-		];
-
-		SkinHooks::onSkinTemplateNavigation( $sktemplate, $links );
-
-		$this->assertArrayNotHasKey( 'userpage', $links['user-menu'] );
-		$this->assertArrayHasKey( 'preferences', $links['user-menu'] );
-	}
-
-	public function testUserMenuRemovesTmpuserpageAndUserpageForTemp(): void {
-		$sktemplate = $this->createSkinTemplateWithUser( true, true );
-		$links = [
-			'user-menu' => [
-				'tmpuserpage' => [ 'id' => 'pt-tmpuserpage' ],
-				'userpage' => [ 'id' => 'pt-userpage' ],
-				'mytalk' => [ 'id' => 'pt-mytalk' ],
-			],
-		];
-
-		SkinHooks::onSkinTemplateNavigation( $sktemplate, $links );
-
-		$this->assertArrayNotHasKey( 'tmpuserpage', $links['user-menu'] );
-		$this->assertArrayNotHasKey( 'userpage', $links['user-menu'] );
-		$this->assertArrayHasKey( 'mytalk', $links['user-menu'] );
-	}
-
-	public function testUserMenuMovesAccountLinksBeforeLogoutForTemp(): void {
-		$sktemplate = $this->createSkinTemplateWithUser( true, true );
-		$links = [
-			'user-menu' => [
-				'userpage' => [ 'id' => 'pt-userpage' ],
-				'createaccount' => [ 'id' => 'pt-createaccount' ],
-				'login' => [ 'id' => 'pt-login' ],
-				'mytalk' => [ 'id' => 'pt-mytalk' ],
-				'mycontris' => [ 'id' => 'pt-mycontris' ],
-				'logout' => [ 'id' => 'pt-logout' ],
-			],
-		];
-
-		SkinHooks::onSkinTemplateNavigation( $sktemplate, $links );
-
-		$this->assertSame(
-			[ 'mytalk', 'mycontris', 'createaccount', 'login', 'logout' ],
-			array_keys( $links['user-menu'] )
-		);
-	}
-
-	public function testUserMenuRemovesAnonuserpageForAnon(): void {
-		$sktemplate = $this->createSkinTemplateWithUser( false, false );
-		$links = [
-			'user-menu' => [
-				'anonuserpage' => [ 'id' => 'pt-anonuserpage' ],
-				'login' => [ 'id' => 'pt-login' ],
-			],
-		];
-
-		SkinHooks::onSkinTemplateNavigation( $sktemplate, $links );
-
-		$this->assertArrayNotHasKey( 'anonuserpage', $links['user-menu'] );
-		$this->assertArrayHasKey( 'login', $links['user-menu'] );
-	}
-
 	public function testToolboxMenuRemovesSiteTools(): void {
 		$skinMock = $this->createMock( Skin::class );
 		$skinMock->method( 'getSkinName' )->willReturn( 'citizen' );
@@ -339,212 +129,6 @@ class SkinHooksTest extends MediaWikiIntegrationTestCase {
 
 		$this->assertSame( 'printer', $sidebar['TOOLBOX']['print']['icon'] );
 		$this->assertSame( 'recentChanges', $sidebar['TOOLBOX']['recentchangeslinked']['icon'] );
-	}
-
-	public function testViewsMenuAddsCdxButtonClasses(): void {
-		$sktemplate = $this->createSkinTemplateMock();
-		$links = [
-			'views' => [
-				'view' => [ 'id' => 'ca-view' ],
-				'history' => [ 'id' => 'ca-history' ],
-			],
-		];
-
-		SkinHooks::onSkinTemplateNavigation( $sktemplate, $links );
-
-		$expectedClasses = [
-			'citizen-cdx-button--size-large',
-			'cdx-button',
-			'cdx-button--fake-button',
-			'cdx-button--fake-button--enabled',
-			'cdx-button--weight-quiet',
-		];
-		foreach ( $expectedClasses as $class ) {
-			$this->assertContains( $class, $links['views']['view']['link-class'] );
-			$this->assertContains( $class, $links['views']['history']['link-class'] );
-		}
-	}
-
-	public function testViewsMenuPromotesEditButtonsToProgressive(): void {
-		$sktemplate = $this->createSkinTemplateMock();
-		$links = [
-			'views' => [
-				'edit' => [ 'id' => 'ca-edit', 'class' => '' ],
-			],
-		];
-
-		SkinHooks::onSkinTemplateNavigation( $sktemplate, $links );
-
-		$linkClass = $links['views']['edit']['link-class'];
-		$this->assertContains( 'cdx-button--weight-primary', $linkClass );
-		$this->assertContains( 'cdx-button--action-progressive', $linkClass );
-		$this->assertNotContains( 'cdx-button--weight-quiet', $linkClass );
-	}
-
-	public function testAssociatedPagesMenuAddsCdxButtonClasses(): void {
-		$sktemplate = $this->createSkinTemplateMock();
-		$links = [
-			'associated-pages' => [
-				'main' => [ 'id' => 'ca-nstab-main' ],
-			],
-		];
-
-		SkinHooks::onSkinTemplateNavigation( $sktemplate, $links );
-
-		$this->assertContains( 'cdx-button', $links['associated-pages']['main']['link-class'] );
-		$this->assertContains( 'citizen-cdx-button--size-large', $links['associated-pages']['main']['link-class'] );
-	}
-
-	/**
-	 * Build a SkinCitizen mock whose setNotificationData() records the merged
-	 * data into $captured, so the notifications tests can assert what the hook
-	 * captured.
-	 *
-	 * @param ?array &$captured Receives the data passed to setNotificationData
-	 * @return SkinCitizen
-	 */
-	private function createCitizenMockCapturing( ?array &$captured ): SkinCitizen {
-		$skin = $this->createMock( SkinCitizen::class );
-		$skin->method( 'getSkinName' )->willReturn( 'citizen' );
-		$skin->method( 'setNotificationData' )->willReturnCallback(
-			static function ( array $data ) use ( &$captured ): void {
-				$captured = $data;
-			}
-		);
-		return $skin;
-	}
-
-	public function testNotificationsMenuCapturesMergedDataAndDropsPortlet(): void {
-		$captured = null;
-		$skin = $this->createCitizenMockCapturing( $captured );
-		$links = [
-			'notifications' => [
-				'notifications-alert' => [
-					'id' => 'pt-notifications-alert',
-					'href' => '/wiki/Special:Notifications',
-					'data' => [ 'counter-num' => 3 ],
-				],
-				'notifications-notice' => [
-					'id' => 'pt-notifications-notice',
-					'href' => '/wiki/Special:Notifications',
-					'data' => [ 'counter-num' => 2 ],
-				],
-			],
-		];
-
-		SkinHooks::onSkinTemplateNavigation( $skin, $links );
-
-		// Echo's portlet is dropped; Citizen renders its own dropdown.
-		$this->assertArrayNotHasKey( 'notifications', $links );
-
-		// The merged data is handed to the skin: combined count (alert + notice)
-		// and the Special:Notifications target.
-		$this->assertSame( 5, $captured['count'] );
-		$this->assertSame( '/wiki/Special:Notifications', $captured['href'] );
-
-		// The panel's pre-mount placeholder previews what is waiting.
-		$this->assertTrue( $captured['has-unread'] );
-		$this->assertCount( 5, $captured['placeholder-rows'] );
-		$this->assertStringContainsString( '#mw-prefsection-echo', $captured['prefs-href'] );
-	}
-
-	public function testNotificationsPlaceholderPreviewsNothingWhenNothingIsWaiting(): void {
-		$captured = null;
-		$skin = $this->createCitizenMockCapturing( $captured );
-		// Echo still renders its badges at zero; the panel has nothing to preview.
-		$links = [
-			'notifications' => [
-				'notifications-alert' => [
-					'id' => 'pt-notifications-alert',
-					'href' => '/wiki/Special:Notifications',
-					'data' => [ 'counter-num' => 0 ],
-				],
-			],
-		];
-
-		SkinHooks::onSkinTemplateNavigation( $skin, $links );
-
-		// No rows and has-unread false, so the template renders the empty state
-		// outright rather than a shimmer that resolves into it.
-		$this->assertSame( 0, $captured['count'] );
-		$this->assertFalse( $captured['has-unread'] );
-		$this->assertSame( [], $captured['placeholder-rows'] );
-	}
-
-	public function testNotificationsPlaceholderCapsItsPreviewAtAPanelful(): void {
-		$captured = null;
-		$skin = $this->createCitizenMockCapturing( $captured );
-		$links = [
-			'notifications' => [
-				'notifications-alert' => [
-					'id' => 'pt-notifications-alert',
-					'href' => '/wiki/Special:Notifications',
-					'data' => [ 'counter-num' => 40 ],
-				],
-			],
-		];
-
-		SkinHooks::onSkinTemplateNavigation( $skin, $links );
-
-		// The preview fills the panel; the list itself scrolls once mounted.
-		$this->assertSame( 40, $captured['count'] );
-		$this->assertCount( 5, $captured['placeholder-rows'] );
-	}
-
-	public function testNotificationsPlaceholderPreviewsOneRowPerNotification(): void {
-		$captured = null;
-		$skin = $this->createCitizenMockCapturing( $captured );
-		$links = [
-			'notifications' => [
-				'notifications-alert' => [
-					'id' => 'pt-notifications-alert',
-					'href' => '/wiki/Special:Notifications',
-					'data' => [ 'counter-num' => 2 ],
-				],
-			],
-		];
-
-		SkinHooks::onSkinTemplateNavigation( $skin, $links );
-
-		// Below the cap the preview matches the count, so two notifications do
-		// not look like five on the way in.
-		$this->assertCount( 2, $captured['placeholder-rows'] );
-	}
-
-	public function testNotificationsMenuLeavesEmptyPortletAlone(): void {
-		$captured = null;
-		$skin = $this->createCitizenMockCapturing( $captured );
-		// Echo's portlet is present but provided neither badge.
-		$links = [ 'notifications' => [] ];
-
-		SkinHooks::onSkinTemplateNavigation( $skin, $links );
-
-		// Nothing to merge: the skin gets no data and the empty portlet is left
-		// untouched (not dropped).
-		$this->assertNull( $captured );
-		$this->assertArrayHasKey( 'notifications', $links );
-	}
-
-	public function testNotificationsMenuWithOnlyNoticeBadge(): void {
-		$captured = null;
-		$skin = $this->createCitizenMockCapturing( $captured );
-		$links = [
-			'notifications' => [
-				'notifications-notice' => [
-					'id' => 'pt-notifications-notice',
-					'href' => '/wiki/Special:Notifications?notice',
-					'data' => [ 'counter-num' => 2 ],
-				],
-			],
-		];
-
-		SkinHooks::onSkinTemplateNavigation( $skin, $links );
-
-		// A single badge is enough: count is the notice count, and the href
-		// falls back to the notice badge's target.
-		$this->assertSame( 2, $captured['count'] );
-		$this->assertSame( '/wiki/Special:Notifications?notice', $captured['href'] );
-		$this->assertArrayNotHasKey( 'notifications', $links );
 	}
 
 	/**
