+++ tests/phpunit/integration/SkinCitizenPreferenceDefaultsTest.php
@@ -40,6 +40,7 @@ class SkinCitizenPreferenceDefaultsTest extends MediaWikiIntegrationTestCase {
 			$this->getServiceContainer()->getUserGroupManager(),
 			$this->getServiceContainer()->getUrlUtils(),
 			$this->getServiceContainer()->getTempUserConfig(),
+			$this->getServiceContainer()->getSpecialPageFactory(),
 			null,
 			[
 				'name' => 'Citizen',
